fix: update return types from Float32 to Int64 for COUNT functions and related tests

+++ src/DatumIngest/Functions/FunctionDocumentation.cs
@@ -1930,7 +1930,7 @@ public static class FunctionDocumentation
         {
             Name = "COUNT",
             Parameters = [Parameter("expression", "Any", isOptional: true)],
-            ReturnType = "Float32",
+            ReturnType = "Int64",
             Description = "Counts the number of rows. COUNT(*) counts all rows; COUNT(expr) counts non-null values.",
             IsAggregate = true,
             Category = FunctionCategory.Aggregate,
