+++ tests/DatumIngest.Tests/Execution/ParallelOperatorTests.cs
@@ -273,7 +273,7 @@ public sealed class ParallelOperatorTests
         List<Row> results = await CollectAsync(groupBy, context);
 
         Assert.Single(results);
-        Assert.Equal(3f, results[0]["COUNT(*)"].AsFloat32());
+        Assert.Equal(3L, results[0]["COUNT(*)"].AsInt64());
     }
 
     /// <summary>
@@ -345,10 +345,10 @@ public sealed class ParallelOperatorTests
         Row groupA = results.First(row => row["category"].AsString() == "A");
         Row groupB = results.First(row => row["category"].AsString() == "B");
 
-        Assert.Equal(3f, groupA["COUNT(*)"].AsFloat32());
+        Assert.Equal(3L, groupA["COUNT(*)"].AsInt64());
         Assert.Equal(9f, groupA["SUM(value)"].AsFloat32());
 
-        Assert.Equal(2f, groupB["COUNT(*)"].AsFloat32());
+        Assert.Equal(2L, groupB["COUNT(*)"].AsInt64());
         Assert.Equal(6f, groupB["SUM(value)"].AsFloat32());
     }
 
@@ -383,12 +383,12 @@ public sealed class ParallelOperatorTests
 
         Row groupXActive = results.First(row =>
             row["dept"].AsString() == "X" && row["status"].AsString() == "active");
-        Assert.Equal(2f, groupXActive["COUNT(*)"].AsFloat32());
+        Assert.Equal(2L, groupXActive["COUNT(*)"].AsInt64());
         Assert.Equal(400f, groupXActive["SUM(amount)"].AsFloat32());
 
         Row groupXInactive = results.First(row =>
             row["dept"].AsString() == "X" && row["status"].AsString() == "inactive");
-        Assert.Equal(1f, groupXInactive["COUNT(*)"].AsFloat32());
+        Assert.Equal(1L, groupXInactive["COUNT(*)"].AsInt64());
         Assert.Equal(200f, groupXInactive["SUM(amount)"].AsFloat32());
     }
 
@@ -470,7 +470,7 @@ public sealed class ParallelOperatorTests
         List<Row> results = await CollectAsync(groupBy, context);
 
         Assert.Single(results);
-        Assert.Equal(0f, results[0]["COUNT(*)"].AsFloat32());
+        Assert.Equal(0L, results[0]["COUNT(*)"].AsInt64());
     }
 
     /// <summary>
@@ -498,10 +498,10 @@ public sealed class ParallelOperatorTests
         Assert.Equal(2, results.Count);
 
         Row nullGroup = results.First(row => row["category"].IsNull);
-        Assert.Equal(2f, nullGroup["COUNT(*)"].AsFloat32());
+        Assert.Equal(2L, nullGroup["COUNT(*)"].AsInt64());
 
         Row groupA = results.First(row => !row["category"].IsNull);
-        Assert.Equal(1f, groupA["COUNT(*)"].AsFloat32());
+        Assert.Equal(1L, groupA["COUNT(*)"].AsInt64());
     }
 
     /// <summary>
@@ -539,7 +539,7 @@ public sealed class ParallelOperatorTests
         // Each group has exactly 10 rows.
         foreach (Row result in results)
         {
-            Assert.Equal(10f, result["COUNT(*)"].AsFloat32());
+            Assert.Equal(10L, result["COUNT(*)"].AsInt64());
         }
 
         // Total sum of all values: 0+1+...+99 = 4950.
@@ -676,7 +676,7 @@ public sealed class ParallelOperatorTests
         // or simply not be merged — both would produce a count other than 10.
         foreach (Row result in results)
         {
-            Assert.Equal(10f, result["COUNT(*)"].AsFloat32());
+            Assert.Equal(10L, result["COUNT(*)"].AsInt64());
             Assert.Equal(10f, result["SUM(value)"].AsFloat32());
         }
     }
@@ -723,10 +723,10 @@ public sealed class ParallelOperatorTests
             string groupName = result["group"].AsString();
             int groupIndex = int.Parse(groupName.AsSpan(1));
 
-            float expectedCount = 4f;
+            long expectedCount = 4L;
             float expectedSum = groupIndex + (groupIndex + 100) + (groupIndex + 200) + (groupIndex + 300);
 
-            Assert.Equal(expectedCount, result["COUNT(*)"].AsFloat32());
+            Assert.Equal(expectedCount, result["COUNT(*)"].AsInt64());
             Assert.Equal(expectedSum, result["SUM(value)"].AsFloat32());
         }
     }
